pyproject.toml
[project]
name = "graphql-voting-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.1",
    "apscheduler>=3.11.1",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.119.1",
    "graphene>=3.0",
    "graphene-sqlalchemy>=2.1.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.1.1",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "starlette-graphene3>=0.6.0",
]
